+++ aut/__init__.py
@@ -1,0 +1,6 @@
+"""
+Top-level module for aut functions used by the functions in
+aut.commands that are called when 'aut' is used on the
+command-line.
+"""
+__version__ = "0.0.1"