+++ autcli/__init__.py
@@ -1,6 +1,0 @@
-"""
-Top-level module for autcli functions used by the functions in
-autcli.commands that are called when 'aut' is used on the
-command-line.
-"""
-__version__ = "0.0.1"